Big error handling fix
moved load 'plugin' inside of error handling block; now, if there's a syntax
error in the loaded file that gets run during the load, the bot core *wont*
crash. This is a Good Thing(tm).

added error handler wrapper method #meth_wrap_proc which takes a
command-method as an argument, wraps it in an error handler and returns a proc
object to be registered with Isaac::Bot's event handler. This prevents errors
in plugins from crashing the Bot core.