Added the ability to 'hide' methods so they don't become callable commands.
The concept of Isaac::Bot helpers (helper methods) was to inject methods inside the bot that can be used within events, but aren't themselves events.

This feature in RubOt allows methods whose names begin with an underscore character to be hidden.  Additionally a new context, :helper, has been added.  methods defined in a :helper context are also hidden